Added requests + image storage

+++ app/Http/Controllers/ProductController.php
@@ -4,9 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use App\Http\Requests\StoreUpdateProductRequest;
 
 class ProductController extends Controller
 {
+  protected $request, $user;
+  protected $repository;
+
+  public function __construct (Request $request, Product $product) {
+    $this->repository = $product;
+    $this->request = $request;
+  }
+
 /**
  * Display a listing of the resource.
  *
@@ -33,19 +43,19 @@ class ProductController extends Controller
 /**
  * Store a newly created resource in storage.
  *
- * @param  \Illuminate\Http\Request  $request
+ * @param  App\Http\Requests\StoreUpdateProductRequest  $request
  * @return \Illuminate\Http\Response
  */
-  public function store(Request $request)
-
+  public function store(StoreUpdateProductRequest $request)
   {
-    $request->validate([
-      'name' => 'required',
-      'price' => 'required',
-      'description' => 'required',
-    ]);
+    $data = $request->only('name', 'description', 'price');
 
-    Product::create($request->all());
+    if ($request->hasFile('image') && $request->image->isValid()) {
+      $imagePath = $request->image->store('public/products');
+      $data['image'] = $imagePath;
+    }
+
+    $this->repository->create($data);
     return redirect()->route('products.index')->with('success','Product created successfully.');
   }
 
@@ -56,9 +66,11 @@ class ProductController extends Controller
  * @param  \App\Product  $product
  * @return \Illuminate\Http\Response
  */
-  public function show(Product $product)
-
+  public function show($id)
   {
+    if (!$product = $this->repository->find($id)) {
+      return redirect()->back();
+    }
     return view('products.show',compact('product'));
   }
 
@@ -66,12 +78,15 @@ class ProductController extends Controller
 /**
  * Show the form for editing the specified resource.
  *
- * @param  \App\Product  $product
+ * @param  int $id
  * @return \Illuminate\Http\Response
  */
-  public function edit(Product $product)
-
+  public function edit($id)
   {
+    if (!$product = $this->repository->find($id)) {
+      return redirect()->back();
+    }
+
     return view('products.edit',compact('product'));
   }
 
@@ -79,19 +94,28 @@ class ProductController extends Controller
 /**
  * Update the specified resource in storage.
  *
- * @param  \Illuminate\Http\Request  $request
+ * @param  App\Http\Requests\StoreUpdateProductRequest  $request $id
  * @param  \App\Product  $product
  * @return \Illuminate\Http\Response
  */
-  public function update(Request $request, Product $product)
+  public function update(StoreUpdateProductRequest $request, $id)
   {
-    $request->validate([
-        'name' => 'nullable',
-        'price' => 'nullable',
-        'description' => 'nullable',
-    ]);
+    if (!$product = $this->repository->find($id)) {
+      return redirect()->back();
+    }
 
-    $product->update($request->all());
+    $data = $request->all();
+
+    if ($request->hasFile('image') && $request->image->isValid()) {
+      if ($product->image && Storage::exists($product->image)) {
+        Storage::delete($product->image);
+      }
+
+      $imagePath = $request->image->store('public/products');
+      $data['image'] = $imagePath;
+    }
+
+    $product->update($data);
 
     return redirect()->route('products.index')->with('warning','Product updated successfully');
   }
@@ -100,11 +124,14 @@ class ProductController extends Controller
 /**
  * Remove the specified resource from storage.
  *
- * @param  \App\Product  $product
+ * @param  int $id
  * @return \Illuminate\Http\Response
  */
-  public function destroy(Product $product)
+  public function destroy($id)
   {
+    if (!$product = $this->repository->find($id)) {
+      return redirect()->back();
+    }
     $product->delete();
     return redirect()->route('products.index')->with('danger','Product deleted successfully');
   }
